[refactor] config.json .env로 환경변수화 .js로 확장자변경으로 다른 코드 로직 수정,

+++ services/blockedExtension.service.js
@@ -1,24 +1,47 @@
 const blockedExtensionRepository = require("../repositories/blockedExtension.repository");
-
 class BlockedExtensionService {
   getAllExtensions = async () => {
-    return await blockedExtensionRepository.getAll();
+    try {
+      return await blockedExtensionRepository.getAll();
+    } catch (err) {
+      console.error(err);
+      throw err;
+    }
   };
 
   addExtension = async (name) => {
-    const existingExtension = await blockedExtensionRepository.getOne(name);
-    if (existingExtension) {
-      throw new Error('이미 등록된 확장자');
+    try {
+      const existingExtension = await blockedExtensionRepository.getOne(name);
+      if (existingExtension) {
+        throw new Error("이미 등록된 확장자");
+      }
+      return await blockedExtensionRepository.add(name);
+    } catch (err) {
+      console.error(err);
+      throw err;
     }
-    return await blockedExtensionRepository.add(name);
   };
 
   getOneExtension = async (name) => {
-    return await blockedExtensionRepository.getOne(name);
+    try {
+      return await blockedExtensionRepository.getOne(name);
+    } catch (err) {
+      console.error(err);
+      throw err;
+    }
   };
 
   deleteExtension = async (name) => {
-    return await blockedExtensionRepository.delete(name);
+    try {
+      const extension = await this.getOneExtension(name);
+      if (!extension) {
+        throw new Error("해당 확장자가 존재하지 않습니다.");
+      }
+      return await blockedExtensionRepository.delete(name);
+    } catch (err) {
+      console.error(err);
+      throw err;
+    }
   };
 }
